+++ test/files/run/reflection-fieldmirror-accessorsareokay.check
@@ -1,0 +1,6 @@
+true
+42
+2
+true
+2
+2
