+++ test/files/run/reflection-fieldmirror-ctorparam.check
@@ -1,1 +1,3 @@
-class java.lang.NoSuchFieldException: Test$A$$x
+class java.lang.Error: this Scala field isn't represented as a Java field, neither it has a Java accessor method
+note that private parameters of class constructors don't get mapped onto fields and/or accessors,
+unless they are used outside of their declaring constructors.
